requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.1=pyhd8ed1ab_1
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py37h540881e_2
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pyh71513ae_1
babel=2.10.3=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
braceexpand=0.1.7=pypi_0
brotlipy=0.7.0=py37h540881e_1004
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py37h43b0acd_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=pypi_0
codecov=2.1.13=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
conda=4.14.0=py37h89c1867_0
conda-package-handling=1.8.1=py37h540881e_1
coverage=7.2.7=pypi_0
cryptography=37.0.4=py37h38fbfac_0
cycler=0.11.0=pypi_0
debugpy=1.6.3=py37hd23a5d3_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.6=pypi_0
dominate=2.8.0=pypi_0
easydict=1.10=pypi_0
einops=0.6.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.1=pypi_0
filelock=3.12.2=pypi_0
flake8=5.0.4=pypi_0
flatbuffers=23.5.26=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
google-auth=2.19.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.54.2=pypi_0
huggingface-hub=0.16.4=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
icu=70.1=h27087fc_0
idna=3.3=pyhd8ed1ab_0
imageio=2.31.0=pypi_0
importlib-metadata=4.2.0=pypi_0
importlib_metadata=4.11.4=hd8ed1ab_0
importlib_resources=5.9.0=pyhd8ed1ab_0
iniconfig=2.0.0=pypi_0
interrogate=1.5.0=pypi_0
ipykernel=6.15.1=pyh210e3f2_0
ipython=7.33.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pypi_0
json5=0.9.5=pyh9f0ad1d_0
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.14.0=pyhd8ed1ab_0
jupyter_client=7.3.5=pyhd8ed1ab_0
jupyter_core=4.11.1=py37h89c1867_0
jupyter_server=1.18.1=pyhd8ed1ab_0
jupyterlab=3.4.3=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.15.1=pyhd8ed1ab_0
kiwisolver=1.4.4=pypi_0
kornia=0.6.12=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h516909a_0
libnsl=2.0.0=h7f98852_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.39.2=h753d276_1
libstdcxx-ng=12.1.0=ha89aaad_16
libuv=1.44.2=h166bdaf_0
libxml2=2.9.14=h22db469_4
libxslt=1.1.35=h8affb1d_0
libzlib=1.2.12=h166bdaf_2
lightning-utilities=0.8.0=pypi_0
lmdb=1.4.1=pypi_0
lxml=4.9.1=py37h540881e_0
markdown=3.3.4=pypi_0
markupsafe=2.1.1=py37h540881e_1
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pypi_0
mistune=2.0.4=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
nbclassic=0.4.3=pyhd8ed1ab_0
nbclient=0.6.7=pyhd8ed1ab_0
nbconvert=7.0.0=pyhd8ed1ab_0
nbconvert-core=7.0.0=pyhd8ed1ab_0
nbconvert-pandoc=7.0.0=pyhd8ed1ab_0
nbformat=5.4.0=pyhd8ed1ab_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
networkx=2.6.3=pypi_0
nodejs=18.8.0=h96d913c_0
notebook=6.4.12=pyha770c72_0
notebook-shim=0.1.0=pyhd8ed1ab_0
numpy=1.21.6=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.14.1=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1w=hd590300_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pandoc=2.19.2=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.0=pypi_0
pip=22.2.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
pluggy=1.0.0=pypi_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.30=pyha770c72_0
protobuf=3.20.3=pypi_0
psutil=5.9.1=py37h540881e_0
ptyprocess=0.7.0=pyhd3deb0d_0
py=1.11.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycodestyle=2.9.1=pypi_0
pycosat=0.6.3=py37h540881e_1010
pycparser=2.21=pyhd8ed1ab_0
pyflakes=2.5.0=pypi_0
pygments=2.13.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.1=py37h540881e_1
pysocks=1.7.1=py37h89c1867_5
pytest=7.3.1=pypi_0
pytest-runner=6.0.0=pypi_0
python=3.7.12=hb7a2778_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.1=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch-lightning=1.9.5=pypi_0
pytz=2022.2.1=pyhd8ed1ab_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.1=py37h0c0c2a8_0
qudida=0.0.4=pypi_0
readline=8.1.2=h0f457ee_0
requests=2.28.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
ruamel_yaml=0.15.80=py37h540881e_1007
safetensors=0.3.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.3.0=py37h89c1867_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py37h89c1867_3
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.39.2=h4ff8645_1
sympy=1.10.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.15.0=py37h89c1867_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.9.5=pypi_0
tini=0.18.0=h14c3975_1001
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pyhd8ed1ab_0
torch=1.12.1+cu116=pypi_0
torchaudio=0.12.1+cu116=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.13.1+cu116=pypi_0
tornado=6.2=py37h540881e_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.3.0=pyhd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
visdom=0.2.4=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webdataset=0.2.48=pypi_0
webencodings=0.5.1=py_1
websocket-client=1.4.0=pyhd8ed1ab_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yapf=0.33.0=pypi_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h166bdaf_2
